pyproject.toml
[project]
name = "WindowPositioner"
version = "0.3.0"
description = "Will let you control size, position, always-on-top and can remove titlebar for windowed applications."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "messagebox>=0.1.0",
    "pillow>=11.2.1",
    "pygetwindow>=0.0.9",
    "pyinstaller>=6.14.0",
    "pywin32>=310",
    "pywinstyles>=1.8",
    "requests>=2.32.4",
]
